Because AdventureResult uses the item count for meat counts, we probably want to make it long too... oh dear

+++ src/net/sourceforge/kolmafia/persistence/MallPriceDatabase.java
@@ -144,7 +144,7 @@ public class MallPriceDatabase {
     }
   }
 
-  public static void recordPrice(int itemId, int price, boolean deferred) {
+  public static void recordPrice(int itemId, long price, boolean deferred) {
     long timestamp = MallPriceManager.currentTimeMillis() / 1000L;
     Price p = MallPriceDatabase.prices.get(itemId);
     if (p == null) {
